+++ cmake_setup/cpp_cuda/simpleAddGPU.cu
@@ -1,6 +1,6 @@
 #include "simpleAddGPU.h"
 #include <stdio.h>
-__global__ void randomize_cuda(double* a,double* b, double* c,long N){
+__global__ void randomize_cuda(float* a,float* b, float* c,long N){
     int index = blockIdx.x * blockDim.x + threadIdx.x;
     
     if(index<N){
@@ -10,7 +10,7 @@ __global__ void randomize_cuda(double* a,double* b, double* c,long N){
     }
 };
 
-__global__ void add_cuda(double* a,double* b,double* c,long N){
+__global__ void add_cuda(float* a,float* b,float* c,long N){
     
     int index = blockIdx.x * blockDim.x + threadIdx.x;
     if(index == 2){
